pyproject.toml
[project]
name = "sanpra_easypay"
authors = [
    { name = "Sanpra Software Solutions", email = "sanprasoftwares@gmail.com"}
]
description = "Payments on your fingertips"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    # "frappe~=15.0.0" # Installed and managed by bench.
    "requests",  # For making HTTP requests
    "pycryptodome",  # For cryptographic functionalities (AES, RSA, etc.)
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"


